Add a workaround for properly rebuilding xfbins with unsupported chunks
All of the chunks referenced in the page will be stored in the NuccChunk object, which will then be added to the dictionary on rebuilding.

This required changing the functionality of NuccChunk.init_data() and therefore removing it from all supported chunks, and adding its previous functionality to each one.